fixing bug in the details ticket page

+++ resources/views/tickets/details.blade.php
@@ -29,18 +29,20 @@
                 <span class="label label-info">{{ $user->name }}</span>
             @endforeach
             </p>
-            @if ( ! currentUser()->hasVoted($ticket))
-            {!! Form::open(['route' => ['votes.submit', $ticket->id], 'method' => 'POST']) !!}
-                <button type="submit" class="btn btn-primary">
-                    <span class="glyphicon glyphicon-thumbs-up"></span> Votar
+            @if (Auth::check())
+                @if ( ! currentUser()->hasVoted($ticket))
+                {!! Form::open(['route' => ['votes.submit', $ticket->id], 'method' => 'POST']) !!}
+                    <button type="submit" class="btn btn-primary">
+                        <span class="glyphicon glyphicon-thumbs-up"></span> Votar
+                    </button>
+                {!! Form::close() !!}
+                @else
+                {!! Form::open(['route' => ['votes.destroy', $ticket->id], 'method' => 'DELETE']) !!}
+                <button type="submit" class="btn btn-primary" >
+                    <span class="glyphicon glyphicon-thumbs-down"></span> Quitar voto
                 </button>
-            {!! Form::close() !!}
-            @else
-            {!! Form::open(['route' => ['votes.destroy', $ticket->id], 'method' => 'DELETE']) !!}
-            <button type="submit" class="btn btn-primary" >
-                <span class="glyphicon glyphicon-thumbs-down"></span> Quitar voto
-            </button>
-            {!! Form::close() !!}
+                {!! Form::close() !!}
+                @endif
             @endif
 
             <h3>Nuevo Comentario</h3>
